+++ src/assets/src/vue/triggers.js
@@ -1,5 +1,0 @@
-import { createApp } from 'vue'
-import { store } from './triggerStore.js';
-
-app.use(store);
-app.mount('#main');