+++ src/events/EmailEvent.php
@@ -2,11 +2,19 @@
 
 namespace Ryssbowh\CraftEmails\events;
 
+use Ryssbowh\CraftEmails\models\Email;
+use Ryssbowh\CraftEmails\records\Email as EmailRecord;
 use yii\base\Event;
 
 class EmailEvent extends Event
 {
+    /**
+     * @var Email|EmailRecord
+     */
     public $email;
 
+    /**
+     * @var bool
+     */
     public $isNew;
 }