+++ src/records/Email.php
@@ -7,11 +7,19 @@ use craft\db\ActiveRecord;
 
 class Email extends ActiveRecord
 {
+    /**
+     * @inheritDoc
+     */
     public static function tableName()
     {
         return '{{%emails}}';
     }
 
+    /**
+     * Turn record to model
+     * 
+     * @return EmailModel
+     */
     public function toModel()
     {
         return new EmailModel($this->getAttributes());