+++ src/records/EmailAttachement.php
@@ -7,11 +7,17 @@ use craft\records\SystemMessage;
 
 class EmailAttachement extends ActiveRecord
 {
+    /**
+     * @inheritDoc
+     */
     public static function tableName()
     {
         return '{{%emails_attachements}}';
     }
 
+    /**
+     * Get system message foreign record
+     */
     public function getSystemMessage()
     {
         return $this->hasOne(SystemMessage::class, ['id' => 'message_id']);